Rename Platforms/OpenMP to Platforms/OMPTarget.
