+++ src/Estimators/tests/CMakeLists.txt
@@ -22,7 +22,7 @@ SET(SRCS test_accumulator.cpp test_local_energy_est.cpp EstimatorManagerBaseTest
 ADD_EXECUTABLE(${UTEST_EXE} ${SRCS})
 TARGET_LINK_LIBRARIES(${UTEST_EXE} catch_main qmcestimators_unit)
 IF(USE_OBJECT_TARGET)
-TARGET_LINK_LIBRARIES(${UTEST_EXE} qmcestimators_unit qmcham_unit qmcwfs qmcparticle qmcutil containers platform_omp)
+TARGET_LINK_LIBRARIES(${UTEST_EXE} qmcestimators_unit qmcham_unit qmcwfs qmcparticle qmcutil containers platform_omptarget)
 ENDIF()
 
 ADD_UNIT_TEST(${UTEST_NAME} "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
@@ -34,7 +34,7 @@ IF(HAVE_MPI)
   SET(SRCS EstimatorManagerNewTest.cpp test_manager_mpi.cpp)
   ADD_EXECUTABLE(${UTEST_EXE} ${SRCS})
   IF(USE_OBJECT_TARGET)
-    TARGET_LINK_LIBRARIES(${UTEST_EXE} qmcestimators_unit qmcham_unit qmcdriver_unit qmcwfs qmcparticle qmcutil containers platform_omp)
+    TARGET_LINK_LIBRARIES(${UTEST_EXE} qmcestimators_unit qmcham_unit qmcdriver_unit qmcwfs qmcparticle qmcutil containers platform_omptarget)
   ENDIF()
   TARGET_LINK_LIBRARIES(${UTEST_EXE} catch_main qmcestimators_unit)
   # Right now the unified driver mpi tests are hard coded for 3 MPI ranks
