+++ src/Platforms/CMakeLists.txt
@@ -34,8 +34,8 @@ TARGET_LINK_LIBRARIES(platform_device PUBLIC message)
 SUBDIRS(CPU)
 TARGET_LINK_LIBRARIES(platform_device PUBLIC platform_cpu)
 
-SUBDIRS(OpenMP)
-TARGET_LINK_LIBRARIES(platform_device PRIVATE platform_omp)
+SUBDIRS(OMPTarget)
+TARGET_LINK_LIBRARIES(platform_device PRIVATE platform_omptarget)
 
 IF(QMC_CUDA)
   SUBDIRS(CUDA_legacy)
