+++ src/Platforms/OMPTarget/CMakeLists.txt
@@ -15,11 +15,11 @@ SET(OMP_SRCS
 )
 
 IF(USE_OBJECT_TARGET)
-  ADD_LIBRARY(platform_omp OBJECT ${OMP_SRCS})
+  ADD_LIBRARY(platform_omptarget OBJECT ${OMP_SRCS})
 ELSE()
-  ADD_LIBRARY(platform_omp ${OMP_SRCS})
+  ADD_LIBRARY(platform_omptarget ${OMP_SRCS})
 ENDIF()
 
 IF(NOT QMC_COMPLEX)
-  TARGET_COMPILE_DEFINITIONS(platform_omp PRIVATE -DOPENMP_NO_COMPLEX)
+  TARGET_COMPILE_DEFINITIONS(platform_omptarget PRIVATE -DOPENMP_NO_COMPLEX)
 ENDIF()
