+++ src/Platforms/tests/OpenMP/test_vector.cpp
@@ -15,7 +15,7 @@
 #include <memory>
 #include <vector>
 #include <iostream>
-#include "OpenMP/OMPallocator.hpp"
+#include "OMPTarget/OMPallocator.hpp"
 
 namespace qmcplusplus
 {
