+++ src/QMCApp/tests/CMakeLists.txt
@@ -22,7 +22,7 @@ ADD_EXECUTABLE(${UTEST_EXE} test_project_data.cpp test_rng_control.cpp)
 
 TARGET_LINK_LIBRARIES(${UTEST_EXE} catch_main qmc)
 IF(USE_OBJECT_TARGET)
-  TARGET_LINK_LIBRARIES(${UTEST_EXE} qmcestimators qmcdriver qmcham qmcwfs qmcparticle qmcutil containers platform_device platform_omp)
+  TARGET_LINK_LIBRARIES(${UTEST_EXE} qmcestimators qmcdriver qmcham qmcwfs qmcparticle qmcutil containers platform_device platform_omptarget)
 ENDIF()
 
 ADD_UNIT_TEST(${UTEST_NAME} "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
