+++ src/QMCWaveFunctions/Fermion/MatrixDelayedUpdateCUDA.h
@@ -14,11 +14,11 @@
 
 #include "CPU/SIMD/aligned_allocator.hpp"
 #include "Platforms/PinnedAllocator.h"
-#include "OpenMP/OMPallocator.hpp"
+#include "OMPTarget/OMPallocator.hpp"
 #include "OhmmsPETE/OhmmsVector.h"
 #include "OhmmsPETE/OhmmsMatrix.h"
 #include "QMCWaveFunctions/Fermion/DiracMatrix.h"
-#include "Platforms/OpenMP/ompBLAS.hpp"
+#include "Platforms/OMPTarget/ompBLAS.hpp"
 #include <cuda_runtime_api.h>
 #include "CUDA/cuBLAS.hpp"
 #include "CUDA/cuBLAS_missing_functions.hpp"
